backend/requirements.txt
flask
supabase
pandas
requests
python-dotenv
